+++ src/utilities/validation.ts
@@ -1,6 +1,5 @@
 /**
  * isRecord is a type check for records
- *
  * @param obj - The object to test
  * @returns True if the object is a record, otherwise false
  */
@@ -9,7 +8,6 @@ export const isRecord = (obj: unknown): obj is Record<string, unknown> =>
 
 /**
  * isString is a type check for strings
- *
  * @param obj - The object to test
  * @returns True if the object is a string, otherwise false
  */
@@ -17,7 +15,6 @@ export const isString = (obj: unknown): obj is string => typeof obj == "string";
 
 /**
  * isNumber is a type check for numbers
- *
  * @param obj - The object to test
  * @returns True if the object is a number, otherwise false
  */
@@ -25,7 +22,6 @@ export const isNumber = (obj: unknown): obj is number => typeof obj == "number";
 
 /**
  * isArrayOfType
- *
  * @param obj - The object to test
  * @param typeCheck - The type check function to validate items in the array
  * @returns True if the object is an array matching the type check, otherwise false